chatbot/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
requests>=2.31
python-dotenv>=1.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
